refactor: 변경된 책임에 맞게 테스트 수정 더 이상 검증하지 않는 케이스에 대한 테스트를 제거
