+++ README.md
@@ -38,7 +38,7 @@
 
 ## 개선 사항
 - [ ] 객체의 책임과 역할에 맞는 검증 로직 갖기
-  - [ ] 데이터 전달의 역할을 담당하는 `DTO` 클래스에서 도메인 객체의 논리적인 특징에 대한 검증 분리
+  - [x] 데이터 전달의 역할을 담당하는 `DTO` 클래스에서 도메인 객체의 논리적인 특징에 대한 검증 분리
 - [ ] 객체의 캡슐화를 유지하며 테스트에 용이한 코드 작성하기
 - [ ] 의미 있는 단위 테스트 고려하기
 - [ ] 역할이 많은 객체 분리하기
